Add files via upload
This is my first ever android application that I've created, although the idea is not entirely mine but I've made some changes to the original application and made this one. What It does is it gives English equivalents of Spanish words and phrases so that a beginner Spanish speaker can get the hang of the language.